pyproject.toml
[project]
name = "oriss_2"
version = "0.1.0"
description = ""
authors = [
    {name = "Kabshah",email = "iamkabshah@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0.0"

dependencies = [
    "langchain-community (>=0.3.30,<0.4.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "langchain-anthropic (>=0.3.21,<0.4.0)",
    "elevenlabs (>=2.16.0,<3.0.0)",
    "ollama (>=0.6.0,<0.7.0)",
    "beautifulsoup4 (>=4.14.2,<5.0.0)",
    "aiolimiter (>=1.2.1,<2.0.0)",
    "lxml (>=6.0.2,<7.0.0)",
    "langchain-mcp-adapters (>=0.1.11,<0.2.0)",
    "fastapi (>=0.118.0,<0.119.0)",
    "langgraph (>=0.6.8,<0.7.0)",
    "mcp (>=1.16.0,<2.0.0)",
    "anthropic (>=0.69.0,<0.70.0)",
    "langchain-groq (>=0.3.8,<0.4.0)",
    "streamlit (>=1.50.0,<2.0.0)",
    "huggingface-hub (>=0.35.3,<0.36.0)",
    "langchain-huggingface (>=0.3.1,<0.4.0)"
]

[tool.poetry]
packages = [{include = "oriss_2"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
